dynamic state source+moved to 64 bit offsets
writeMessage logic was inverted (and data source too)
limit output buffer for consumer output
eof that doesn't break things and actually use id for consuming
subscriber tracking logic implemented
eof offset guards added
surveillance! and named services!
kanban shuts down correctly after kafka error
it's possible now to just acknowledge a stream message pack
offset problem fixed
last id based recovery works
stream should clear outstanding data request when it sent something.
Acknowledge now does this too (although this shouldn't happen unless
clients fires ack for no reason but it's now possible)
flushlater fixed

+++ actors/expected_interfaces.go
@@ -10,9 +10,9 @@ type (
 	}
 
 	StateChangeStream interface {
-		FillData(data inspect.Inspectable, offset int, maxLen int) (result inspect.Inspectable, nextOffset int, err error)
-		LastOffsetChanged(offset int)
-		GetLatestState() (lastOffset int, lastStateSource DataSource)
+		FillData(data inspect.Inspectable, offset int64, maxLen int) (result inspect.Inspectable, nextOffset int64, err error)
+		LastOffsetChanged(offset int64)
+		GetLatestState() (lastOffset int64, lastStateSource DataSource)
 		NoMoreSubscribers()
 	}
 
@@ -22,7 +22,7 @@ type (
 	}
 	DynamicDataSource interface {
 		DataSource
-		RequestNext(onDataAvailable func(), maxOffset int) bool
+		RequestNext(onDataAvailable func(), maxOffset int64) bool
 	}
 )
